Refactor display menu, edit exit display screen
